requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-classic
langchain-text-splitters
langchain-openai
langchain-community
langchain-huggingface
langchain-chroma
chromadb
tiktoken
pypdf
jinja2
sentence-transformers
python-multipart
